% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_3D_to_gif.R
\name{KG_3D_to_gif}
\alias{KG_3D_to_gif}
\title{Create a 3D plot as a GIF}
\usage{
KG_3D_to_gif(
  seurat_object,
  dims = 1:30,
  pt.size = 2,
  cols = NULL,
  group.by = NULL,
  run_umap = TRUE,
  reduction = "pca",
  min.dist = 0.3,
  spread = 1,
  bg_col = "white",
  type = "p",
  decorate = TRUE,
  file = "plot.gif",
  fps = 5,
  trajectory_coords = NULL,
  trajectory_col = NULL
)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{dims}{A vector of dimensions to use.}

\item{pt.size}{Size of data points.}

\item{cols}{A vector of colors to use for the shading of data points.}

\item{group.by}{Group identifier to use for shading.}

\item{run_umap}{Whether or not to run UMAP from scratch.}

\item{reduction}{The dimensionality reduction to use.}

\item{min.dist}{Controls how tightly embeddings are allowed to compress points together. Larger values result in more even distribution. Smaller values optimize more accurately the local structure.}

\item{spread}{Effective scale of embedded points. Determines, along with min.dist, how clumped embedded points are.}

\item{bg_col}{The background color. Currently not supported.}

\item{type}{Same as in rgl::plot3d; for the purpose of scRNA-seq, either "p" (point) or "s" (spehere) would be appropriate.}

\item{decorate}{Same as in rgl::plot3d; set as FALSE if you want to remove axes etc.}

\item{file}{The file name to use for the GIF.}

\item{fps}{Frames per secund; controls the speed at which the plot will rotate.}

\item{trajectory_coords}{A data frame with coordinates for plotting a trajectory.}

\item{trajectory_col}{The color for the trajectory, if coordinates are provided.}
}
\description{
This function generates 3D dimension plots as GIFs.
}
\examples{
KG_3D_to_gif(B_cell_dataset, file = "3D_plot.gif")

}
