% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_3DFeaturePlot.R
\name{KG_3DFeaturePlot}
\alias{KG_3DFeaturePlot}
\title{Feature plots, but in 3D!}
\usage{
KG_3DFeaturePlot(
  seurat_object,
  gene,
  dims = 1:35,
  reduction = "pca",
  assay = "RNA",
  pt.size = 7,
  stroke = 1,
  cols = c("lightgrey", "red"),
  bg_col = "black",
  min.dist = 0.3,
  spread = 1,
  trajectory_coords = NULL,
  trajectory_col = NULL,
  lwd = 2
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{gene}{The gene for which the expression should be plotted.}

\item{dims}{The range of dimensions to use from linear dimensional reduction method.}

\item{reduction}{Name of the dimensional reduction method to use as input, "pca" by default.}

\item{assay}{Name of the Seurat object assay to use.}

\item{pt.size}{Controls the size of plotted points/cells.}

\item{stroke}{Size of border to data points.}

\item{cols}{Determines the colors to use for expression levels. Should be a vector of two colors.}

\item{min.dist}{Controls how tightly embeddings are allowed to compress points together. Larger values result in more even distribution. Smaller values optimize more accurately the local structure.}

\item{spread}{Effective scale of embedded points. Determines, along with min.dist, how clumped embedded points are.}

\item{trajectory_coords}{A vector containing coordinates for drawing trajectory lines.}

\item{trajectory_col}{What color to use for the trajectory.}

\item{lwd}{Line width for trajectory.}

\item{bg.col}{Controls the background color of the plot.}

\item{group.by}{Controls how cells should be grouped/colored. By default, uses the default identity of cells in the Seurat object.}
}
\description{
This function generates a feature plot in 3D.
}
\examples{
KG_3DFeaturePlot(B_cell_dataset, gene = "CD27")
}
