% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_3DPlot.R
\name{KG_3DPlot}
\alias{KG_3DPlot}
\title{3D UMAP for Seurat object
This function computes and plots cells in three dimensions.}
\usage{
KG_3DPlot(
  seurat_object,
  dims = 1:35,
  reduction = "pca",
  assay = "RNA",
  pt.size = 7,
  cols = NULL,
  bg_col = "white",
  min.dist = 0.3,
  spread = 1,
  group.by = NULL
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{dims}{The range of dimensions to use from linear dimensional reduction method.}

\item{reduction}{Name of the dimensional reduction method to use as input, "pca" by default.}

\item{assay}{Name of the Seurat object assay to use.}

\item{pt.size}{Controls the size of plotted points/cells.}

\item{cols}{Determines the colors to use for cells. If NULL, uses default colors.}

\item{bg_col}{Controls the background color of the plot.}

\item{min.dist}{Controls how tightly embeddings are allowed to compress points together. Larger values result in more even distribution. Smaller values optimize more accurately the local structure.}

\item{spread}{Effective scale of embedded points. Determines, along with min.dist, how clumped embedded points are.}

\item{group.by}{Controls how cells should be grouped/colored. By default, uses the default identity of cells in the Seurat object.}
}
\description{
3D UMAP for Seurat object
This function computes and plots cells in three dimensions.
}
\examples{
KG_3DPlot(B_cell_dataset)
}
