% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_add_suffix.R
\name{KG_add_suffix}
\alias{KG_add_suffix}
\title{Add suffix to list of barcodes
This function adds a suffix to all cells in the specified input list.}
\usage{
KG_add_suffix(cells, suffix)
}
\arguments{
\item{cells}{A list of cell names.}

\item{suffix}{The suffix to be added to the cell names}
}
\description{
Add suffix to list of barcodes
This function adds a suffix to all cells in the specified input list.
}
\examples{
KG_add_suffix(list_of_cells, "_4")
}
