% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_bs_pal.R
\name{KG_bs_pal}
\alias{KG_bs_pal}
\title{A color palette that swings both ways.}
\usage{
KG_bs_pal(n)
}
\arguments{
\item{n}{The number of colors to generate.}
}
\value{
A character vector of color hex codes.
}
\description{
This function generates a color palette ranging from pink to purple to dark blue.
}
\examples{
KG_bs_pal(50)
}
