% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_clone_tree.R
\name{KG_clone_tree}
\alias{KG_clone_tree}
\title{Construct and plot clonal trees}
\usage{
KG_clone_tree(
  clones,
  tips = "c_call",
  tip_palette = NULL,
  build = "pml",
  tipsize = 7,
  scale = 0,
  layout = "rectangular",
  x_max = NULL,
  breaks = NULL,
  common_scale = FALSE,
  stroke = 0.5
)
}
\arguments{
\item{clones}{An airrClone object from dowser::formatClones.}

\item{tips}{The same as in plotTrees.}

\item{tip_palette}{A vector of colors and names specifying the colors for each state.}

\item{build}{The program used to construct trees (pratchet, pml, dnappars, dnaml, igphyml).}

\item{tipsize}{The size of tree tips.}

\item{scale}{The width of branch length scale bar.}

\item{layout}{The layout of the trees (rectangular, circular, slanted, fan, inward_circular, radial, unrooted, equal_angle, daylight, dendrogram, ape, ellipse, roundrect).}

\item{x_max}{The maximum value of the x-axis. Default is the maximum of the SHM for each specific clone.}

\item{breaks}{The breaks for the x-axis.}

\item{common_scale}{Define if branches should have the same scale across trees.}

\item{stroke}{Thickness of tip border.}
}
\description{
This function uses the dowser and ggtree packages to construct custom clonal trees.
}
\examples{
KG_clone_tree(B_cell_clones)
}
