% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_clones.R
\name{KG_clones}
\alias{KG_clones}
\title{Calling clones in BCR data
This function calls clonotypes in BCR data using the dowser and scoper packages.}
\usage{
KG_clones(bcr_data, threshold = 0.15, reference_dir = NULL)
}
\arguments{
\item{bcr_data}{A data frame containing the BCR data in AIRR format.}

\item{threshold}{Threshold used for hierarchical clustering of sequences. Default is 0.15, i.e., 85\% similarity threhsold.}

\item{reference_dir}{Path to directory with IMGT VDJ references to use.
KG_clones(bcr_dataset)}
}
\description{
Calling clones in BCR data
This function calls clonotypes in BCR data using the dowser and scoper packages.
}
