% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_density.R
\name{KG_density}
\alias{KG_density}
\title{Make nice density plots for feature expression!}
\usage{
KG_density(
  seurat_object,
  genes,
  pal = "viridis",
  pal.option = "viridis",
  pt.size = 3,
  stroke = 0.01,
  joint = FALSE,
  combine = TRUE,
  dims = c(1, 2),
  axes = TRUE,
  legend = TRUE,
  reduction = "umap"
)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{genes}{A gene or vector of genes.}

\item{pal}{Palette used for shading. Supported options are  "viridis", "flame", "ghost", "mutant", "headlight". You can also supply a custom palette as generated by grDevices::colorRampPalette.}

\item{pal.option}{Shading option for viridis palette, same as in viridis::scale_fill_viridis.}

\item{pt.size}{Size of data points.}

\item{stroke}{Size of border to data points. NOTE: when exportet as PDF, borders can alter the appearance of the plot. To avoid this, export externally as PNG file.}

\item{joint}{Logical; whether joint expression should be plotted.}

\item{combine}{Logical; whether plots should be combined or separeate, when genes >1.}

\item{dims}{Dimensions to plot. By default, plots the first two dimensions.}

\item{axes}{Logical; whether axes should be shown.}

\item{legend}{Logical; whether to plot the legend.}

\item{reduction}{The dimensionality reduction to use.}
}
\description{
This functions generates density plots of specified features.
}
\examples{
KG_density(B_cell_data, genes = "CD27")
}
