% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_dimplot.R
\name{KG_dimplot}
\alias{KG_dimplot}
\title{Make a nice dimplot}
\usage{
KG_dimplot(
  seurat_object,
  cells = NULL,
  reduction = NULL,
  dims = c(1, 2),
  group.by = NULL,
  label = FALSE,
  label.size = 4,
  pt.size = 3,
  stroke = 0.2,
  legend = TRUE,
  cols = NULL,
  pub_ready = FALSE
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{cells}{List of cells to plot. If NULL, all cells are plotted.}

\item{reduction}{The dimensionality reduction to use.}

\item{dims}{Dimensions to plot. By default, plots the first two dimensions.}

\item{group.by}{Group identifier to use for shading.}

\item{label}{Whether to plot group identifier labels.}

\item{label.size}{Size of the group identifier labels.}

\item{pt.size}{Size of data points.}

\item{stroke}{Size of border to data points. NOTE: when exportet as PDF, borders can alter the appearance of the plot. To avoid this, export externally as PNG file.}

\item{legend}{Whether or not to plot the legend/key to the group identifier.}

\item{cols}{A vector of colors to use for the shading of data points.}

\item{pub_ready}{Layout option. When TRUE, produces a more clean plot, with shorter axis lines and no ticks.}
}
\description{
This function constructs a nice, clean dimplot from dimensionally reduced data points in a Seurat object.
}
\examples{
KG_dimplto(B_cell_dataset)
}
