% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_dotplot.R
\name{KG_dotplot}
\alias{KG_dotplot}
\title{Dot plot of relative gene expression}
\usage{
KG_dotplot(
  seurat_object,
  genes,
  cols = c("blue4", "red2"),
  border = TRUE,
  title = NULL,
  portrait = TRUE,
  font = "arial"
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{genes}{A vector of genes to plot.}

\item{cols}{A vector of colors to use for low and high values, respectively.}

\item{border}{Whether or not to draw a border to the dots.}

\item{title}{A title for the plot.}

\item{portrait}{Whether or not to plot genes on the y-axis}
}
\description{
This function wraps the Seurat DotPlot function with ggplot2 functions, for added customizability.
}
\examples{
KG_dotplot(B_cell_dataset, genes = c("CD19", "CD27", "CD38", "CR2"))
}
