% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_featureplot.R
\name{KG_featureplot}
\alias{KG_featureplot}
\title{Make a nice feature plot!}
\usage{
KG_featureplot(
  seurat_object,
  genes,
  reduction = "umap",
  dims = c(1, 2),
  pt.size = 2,
  cols = c("#FEFDE2", "#FC8C40", "#950028"),
  stroke = 0.05,
  axes = TRUE,
  legend = TRUE
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{genes}{A gene, or vector of genes, for which expression should be plotted.}

\item{reduction}{The dimensionality reduction to use.}

\item{dims}{Dimensions to plot. By default, plots the first two dimensions.}

\item{pt.size}{Size of data points.}

\item{cols}{A vector of colors to use for the shading of data points. Should be a vector of three colors, for low, mid, and high expression.}

\item{stroke}{Size of border to data points. NOTE: when exportet as PDF, borders can alter the appearance of the plot. To avoid this, export externally as PNG file.}

\item{axes}{Logical; whether axes should be shown.}

\item{legend}{Logical; whether to plot the legend.}

\item{label}{Whether to plot group identifier labels.}

\item{label.size}{Size of the group identifier labels.}
}
\description{
This function constructs nice, clean feature plots from dimensionally reduced data points in a Seurat object.
}
\examples{
KG_featureplot(B_cell_dataset, "CD27")
}
