% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_fetch_genes.R
\name{KG_fetch_genes}
\alias{KG_fetch_genes}
\title{Fetch all genes with specific prefix expressed in the data}
\usage{
KG_fetch_genes(seurat_object, gene_prefix)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{gene_prefix}{The prefix of gene names to be retrieved}
}
\description{
This function fetches all genes with the specified prefix expressed in the dataset
}
\examples{
KG_fetch_genes(pbmc_dataset, "HLA-")
}
