% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_filter_bcr.R
\name{KG_filter_bcr}
\alias{KG_filter_bcr}
\title{Filter AIRR object BCR data
This function filters BCR data output from IgBLAST.}
\usage{
KG_filter_bcr(
  bcr_data,
  productive = TRUE,
  single_IgH = TRUE,
  c_annotated = TRUE,
  hl_pair = TRUE
)
}
\arguments{
\item{bcr_data}{A data frame containing the BCR data in AIRR format.}

\item{productive}{Whether unproductive transcripts should be removed.}

\item{single_IgH}{Whether cells with multiple heavy chains should be removed.}

\item{c_annotated}{Whether cells with no annotated C gene should be removed.}

\item{hl_pair}{Whether cells with no paired heavy/light chain should be removed.}
}
\description{
Filter AIRR object BCR data
This function filters BCR data output from IgBLAST.
}
