% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_filter_pbmcs.R
\name{KG_filter_pbmcs}
\alias{KG_filter_pbmcs}
\title{A function for differentiating B cells, T cells, and monocytes in PBMC scRNA-seq data.}
\usage{
KG_filter_pbmcs(
  seurat_object,
  features = c("CD19", "CD3E", "CD14"),
  scale.factor = 10000
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{features}{The genes to use for cell type determination.}

\item{scale.factor}{Scale factor to use for normalization following filtering.}
}
\description{
A function for differentiating B cells, T cells, and monocytes in PBMC scRNA-seq data.
}
\examples{
KG_filter_pbmcs(B_cell_dataset)
}
