% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_filter_vdj.R
\name{KG_filter_vdj}
\alias{KG_filter_vdj}
\title{Filter V(D)J-seq data.}
\usage{
KG_filter_vdj(
  vdj_data,
  multiple_samples = FALSE,
  remove_duplicate_IGH = TRUE,
  remove_duplicate_IGK = TRUE,
  remove_duplicate_IGL = TRUE,
  remove_no_c = TRUE,
  remove_no_j = TRUE,
  remove_no_v = TRUE,
  productive = TRUE,
  only_paired = TRUE,
  save_filtered = FALSE,
  print_numbers = TRUE
)
}
\arguments{
\item{vdj_data}{A data frame of V(D)J-seq data, or a list of data frames containing multiple V(D)J-seq.}

\item{multiple_samples}{Logical; whether or not vdj_data consists of a single data frame or a list of data frames.}

\item{remove_duplicate_IGH}{Logical; whhether heavy chain duplicates should be removed.}

\item{remove_duplicate_IGK}{Logical; whether kappa light chain duplicates should be removed.}

\item{remove_duplicate_IGL}{Logical; whether lambda light chain duplicates should be removed.}

\item{remove_no_c}{Logical; whether sequences with no annotated C gene should be removed.}

\item{remove_no_j}{Logical; whether sequences with no annotated J gene should be removed.}

\item{remove_no_v}{Logical; whether sequences with no annotated V gene should be removed}

\item{productive}{Logical; whether only productively recombined sequuences should be retained.}

\item{only_paired}{Logical; whether to only keep sequences that have a matching pair of heavy and light chains.}

\item{save_filtered}{Logical; whether sequences tha