% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_flame_pal.R
\name{KG_flame_pal}
\alias{KG_flame_pal}
\title{A flaming hot color palette!}
\usage{
KG_flame_pal(n)
}
\arguments{
\item{n}{The number of colors to generate.}
}
\value{
A character vector of color hex codes.
}
\description{
This function generates a fiery color palette, ranging from black, through red and orange, to gold.
}
\examples{
KG_flame_pal(50)
}
