% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_gene_to_isotype.R
\name{KG_gene_to_isotype}
\alias{KG_gene_to_isotype}
\title{Convert C calls to isotype
This function annotates the major isotypes for each cell in a Seurat object based on the heavy chain c calls from the Cell Ranger output.}
\usage{
KG_gene_to_isotype(seurat_object, c_call)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{c_call}{The column in the Seurat object where the C calls are stored.}
}
\description{
Convert C calls to isotype
This function annotates the major isotypes for each cell in a Seurat object based on the heavy chain c calls from the Cell Ranger output.
}
\examples{
KG_gene_to_isotype(B_cell_dataset, "c_call")
}
