% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_ghost_pal.R
\name{KG_ghost_pal}
\alias{KG_ghost_pal}
\title{An eery color palette!}
\usage{
KG_ghost_pal(n)
}
\arguments{
\item{n}{The number of colors to generate.}
}
\value{
A character vector of color hex codes.
}
\description{
This function generates a color palette from black to successively brighter shades of blue.
}
\examples{
KG_ghost_pal(50)
}
