% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_hc.R
\name{KG_hc}
\alias{KG_hc}
\title{Hierarchical clustering
This function streamlines hierarchical clustering of Seurat clusters.}
\usage{
KG_hc(
  seurat_object,
  genes = NULL,
  assay = "RNA",
  title = NULL,
  ident_col = "seurat_clusters",
  xlim = c(NA, 200)
)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{genes}{A list of genes to base the clustering on. Default is the variable features.}

\item{assay}{The name of the Seurat object assay to use.}

\item{title}{Plot title.}

\item{ident_col}{Identity column to use.Default is 'seurat_clusters'.}

\item{xlim}{Determines x-axis limits.
KG_hc(B_cell_dataset)}
}
\description{
Hierarchical clustering
This function streamlines hierarchical clustering of Seurat clusters.
}
