% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_headlight_pal.R
\name{KG_headlight_pal}
\alias{KG_headlight_pal}
\title{A blinding palette!}
\usage{
KG_headlight_pal(n)
}
\arguments{
\item{n}{The number of colors to generate.}
}
\value{
A character vector of color hex codes.
}
\description{
This function gnereates an illuminating color palette ranging from black, to gold, to white.
}
\examples{
KG_headlight_pal(50)
}
