% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_import_seurat_clusters.R
\name{KG_import_seurat_clusters}
\alias{KG_import_seurat_clusters}
\title{Import Seurat-defined clusters to Monocle object
This function imports the clustering and dimension plot coordinates from a Seurat object and adds them to a Monocle3 object.}
\usage{
KG_import_seurat_clusters(
  monocle_object,
  seurat_object,
  dims = 35,
  plot_3d = FALSE,
  assay = "RNA",
  sparse = FALSE,
  min.dist = 0.3,
  spread = 1,
  ident_col = "seurat_clusters"
)
}
\arguments{
\item{monocle_object}{A monocle object.}

\item{seurat_object}{A Seurat object.}

\item{dims}{The number of principal components to use.}

\item{plot_3d}{Logical, if TRUE, computes and plots in three dimensions.}

\item{assay}{Name of the assay from the Seurat object to use, "RNA" by default.}

\item{sparse}{Logical, if TRUE, attempts to increase distance between cells in the dimension plot.}

\item{min.dist}{If sparse = TRUE, controls how tightly embeddings are allowed to compress points together. Larger values result in more even distribution. Smaller values optimizes more accurately the local structure.}

\item{spread}{Effective scale of embedded points. If sparse = TRUE, determines, along with min.dist, how clumped embedded points are.}

\item{ident_col}{The identity column to use from the Seurat object. By default, 'seurat_clusters' is used.
KG_import_seurat_clusters(monocle_object, B_cell_dataset)}
}
\description{
Import Seurat-defined clusters to Monocle object
This function imports the clustering and dimension plot coordinates from a Seurat object and adds them to a Monocle3 object.
}
