% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_integrate_CCA.R
\name{KG_integrate_CCA}
\alias{KG_integrate_CCA}
\title{Integrate data with Seurat (CCA)}
\usage{
KG_integrate_CCA(
  seurat_objects,
  assay = "SCT",
  nIntegration_features = 10000,
  reduction = "cca",
  dims = 1:35,
  scale_data = TRUE,
  new.name = "Integrated"
)
}
\arguments{
\item{seurat_objects}{A list of Seurat objects to integrate.}

\item{assay}{Name of the assay to use.}

\item{nIntegration_features}{The number of integration features to use.}

\item{reduction}{Name of dimensional reduction method to use in integration}

\item{dims}{The range of principal components to use.}

\item{scale_data}{Logical, whether or not to scale the data following integration.}

\item{new.name}{The name to give to the integrated assay.}
}
\description{
A function for streamlining integration of Seurat objects with Seurat's CCA model.
}
\examples{
KG_integrate_CCA(B_cell_datasets)
}
