% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_isotype_pie.R
\name{KG_isotype_pie}
\alias{KG_isotype_pie}
\title{Plot percentage distribution of isotypes by cluster as pie charts}
\usage{
KG_isotype_pie(
  seurat_object,
  MGA_only = TRUE,
  cluster = NULL,
  combine = TRUE,
  ncol = 2
)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{MGA_only}{Whether or not to only plot IgM, IgG, and IgA isotypes}

\item{cluster}{Name of cluster(s) for which to generate pie charts}

\item{combine}{Whether or not plots should be combined into a single or separate plots}

\item{ncol}{The number of columns for the layout of the plot}
}
\description{
This function plots the distribution of Ig isotypes for clusters in a Seurat object.
}
\examples{
KG_isotype_pie(B_cell_dataset, cluster = "0")
}
