% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_list_pals.R
\name{KG_list_pals}
\alias{KG_list_pals}
\title{List all available color palettes in KG.}
\usage{
KG_list_pals()
}
\value{
A character vector of palette names.
}
\description{
This function lists the names of all color palettes available in this package.
}
\examples{
KG_list_palettes()
}
