% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_matchGEXtoVDJ.R
\name{KG_matchGEXtoVDJ}
\alias{KG_matchGEXtoVDJ}
\title{Match V(D)J-seq sequences to GEX data}
\usage{
KG_matchGEXtoVDJ(
  gex_data,
  vdj_data,
  multiple_samples = FALSE,
  save_filtered = FALSE,
  print_numbers = TRUE
)
}
\arguments{
\item{gex_data}{A Seurat object, or a list of Seurat objects}

\item{vdj_data}{An AIRR data object, or a list of AIRR data objects.}

\item{multiple_samples}{Logical; whether or not vdj_data consists of a single data frame or a list of data frames. NOTE: if set to TRUE, the function expects gex_data is also a list of matching Seurat objects. Matching objects should appear in the same order in both GEX and V(D)J-seq lists.}

\item{save_filtered}{Logical; whether sequences that are removed should be saved to an xlsx file.}

\item{print_numbers}{Logical; whether the number of sequences removed in each step should be printed in the console.}
}
\description{
This function filters V(D)J-seq data based on sequences present in paired GEX data. Sequences that are not present in the GEX data are removed.
}
\examples{
KG_matchGEXtoVDJ(seurat_object, vdj_dataset)

}
