% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_multi_gene_plot.R
\name{KG_multi_gene_plot}
\alias{KG_multi_gene_plot}
\title{View feature or violin plots one by one.}
\usage{
KG_multi_gene_plot(
  seurat_object,
  genes = NULL,
  plot = "feature",
  pt.size = 2,
  feature_cols = c("lightgrey", "red"),
  feature_label = FALSE,
  feature_label_size,
  vln_cols = NULL,
  group.by = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{genes}{A vector of genes to use for plotting.}

\item{plot}{Which kind of plot you want ("feature" or "violin"). Default is "feature".}

\item{pt.size}{Size of dots in feature plots.}

\item{feature_cols}{A vector of colors to use for the feature plot. Default are "lightgrey" and "red".}

\item{feature_label}{Logical; whether or not to plot identity labels on the feature plot.}

\item{feature_label_size}{Size fo the identity labels in the feature plot.}

\item{vln_cols}{A vector of colors to use for identities in the violin plot.}

\item{group.by}{Name of the identity factor in the Seurat object to use. If NULL, the current identity factor is used.}

\item{verbose}{Logical; text or no text?}
}
\description{
This function generates and plots individual feature or violin plots from the Seurat package for each input gene one by one. This is handy when you have a long list of genes and you want to examine the actual expression (as opposed to relative expression in e.g., heatmaps) for each gene, but cannot be bothered to print everything to a PDF file or wrap the plots.
}
\examples{
KG_multi_gene_plot(B_cell_dataset, genes = c("CD19", "MS4A1", "CD27", "CR2"))
}
