% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_mutant_pal.R
\name{KG_mutant_pal}
\alias{KG_mutant_pal}
\title{A slimy color palette for ninja mutatns!}
\usage{
KG_mutant_pal(n)
}
\arguments{
\item{n}{The number of colors to generate.}
}
\value{
A character vector of color hex codes.
}
\description{
This function generates a radioactive color palette, ranging from black through successively brighter shades of green.
}
\examples{
KG_mutant_pal(50)
}
