% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_prep_clones.R
\name{KG_prep_clones}
\alias{KG_prep_clones}
\title{Clonotype analysis}
\usage{
KG_prep_clones(
  vdj_data,
  cell_id = "cell_id",
  only_heavy = TRUE,
  split_light = TRUE,
  summarize_clones = FALSE,
  fields = NULL
)
}
\arguments{
\item{vdj_data}{An AIRR data object.}

\item{cell_id}{Name of the column containing the cell identifiers/barcodes.}

\item{only_heavy}{Logical; whether to only use IGH sequences for grouping.}

\item{split_light}{Logical; whether to split clones for light chains.}

\item{summarize_clones}{Logical; same as in scoper::hierarchicalClones.}

\item{fields}{Vector of additional grouping columns. Sequences with different values will be considered as separate clones. Useful for separation by sample in integrated datasets.}
}
\description{
This function streamlines the functions from the Immcantation framework to define hierarchical clones.
}
\examples{
KG_prep_lones(vdj_data)

}
