% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_princess_pal.R
\name{KG_princess_pal}
\alias{KG_princess_pal}
\title{A sickeningly sweet bubblegum palette!}
\usage{
KG_princess_pal(n)
}
\arguments{
\item{n}{The number of colors to generate.}
}
\value{
A character vector of color hex codes.
}
\description{
This function generates the cotton candy palette of every princess' dream.
}
\examples{
KG_flame_pal(50)
}
