% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_project_3D.R
\name{KG_project_3D}
\alias{KG_project_3D}
\title{Project and plot in 3D
This function projects a Seurat object onto the clusters and UMAP coordinates from a reference, and plots the projected data in 3D.}
\usage{
KG_project_3D(
  seurat_object,
  reference,
  assay = "RNA",
  reference.assay = "RNA",
  reduction = "pca",
  dims = 1:35,
  min.dist = 0.3,
  spread = 1,
  reference.reduction = "pca",
  reduction.model = "umap",
  reference.idents = NULL,
  pt.size = 5,
  cols = NULL,
  bg_col = "white"
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{reference}{A Seurat object from which cluster identities and UMAP coordinates/embeddings should be pulled.}

\item{assay}{Name of the Seurat object assay to use.}

\item{reference.assay}{Name of the Seurat reference object assay to use.}

\item{reduction}{Name of the dimensional reduction method to use as input, "pca" by default.}

\item{dims}{The range of dimensions to use from linear dimensional reduction method.}

\item{min.dist}{Controls how tightly embeddings are allowed to compress points together. Larger values result in more even distribution. Smaller values optimize more accurately the local structure.}

\item{spread}{Effective scale of embedded points. Determines, along with min.dist, how clumped embedded points are.}

\item{reference.reduction}{Name of dimensional reduction method to use as input from reference, "pca" by default.}

\item{reduction.model}{Name of dimensional reduction model to use for plotting, "umap" by default.}

\item{reference.idents}{The identity variable to use from the reference for the projection. By default, uses the default identity of the reference object.}

\item{pt.size}{Controls the size of plotted points/cells.}

\item{cols}{Determines the color palette to use for cells. If NULL, uses default colors.}

\item{bg_col}{Controls the background color of the plot.}
}
\description{
Project and plot in 3D
This function projects a Seurat object onto the clusters and UMAP coordinates from a reference, and plots the projected data in 3D.
}
