% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_rainbow_pal.R
\name{KG_rainbow_pal}
\alias{KG_rainbow_pal}
\title{A prideful rainbow palette!}
\usage{
KG_rainbow_pal(n)
}
\arguments{
\item{n}{The number of colors to generate.}
}
\value{
A character vector of color hex codes.
}
\description{
This function generates an inclusive palette for everyone.
}
\examples{
KG_flame_pal(50)
}
