% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_remove_BCR.R
\name{KG_remove_BCR}
\alias{KG_remove_BCR}
\title{Remove BCR genes}
\usage{
KG_remove_BCR(seurat_object)
}
\arguments{
\item{seurat_object}{A Seurat object
KG_remove_BCR(B_cell_dataset)}
}
\description{
This function removes all B-cell receptor genes from a Seurat object.
}
