% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_subtype_pie.R
\name{KG_subtype_pie}
\alias{KG_subtype_pie}
\title{Plot percentage distribution of isotype subtypes by cluster as pie charts}
\usage{
KG_subtype_pie(
  seurat_object,
  subtype = "IgG",
  c_call = "c_call",
  cluster = NULL,
  combine = TRUE,
  ncol = 2
)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{subtype}{Whether to plot IgG or IgA subtypes}

\item{c_call}{The name of the meta data slot in which the heavy chain C-gene calls are stored}

\item{cluster}{The name(s) of then cluster(s) for which to generate pie charts}

\item{combine}{Whether or not plots should be combined into a single or separate plots}

\item{ncol}{The number of columns for the layout of the plot}
}
\description{
This function plots the distribution of Ig subtypes of IgG (default) or IgA for clusters in a Seurat object.
}
\examples{
KG_subtype_pie(B_cell_dataset, c_call = "c_call", cluster = "0")
}
