% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_vdj_isotypes.R
\name{KG_vdj_isotypes}
\alias{KG_vdj_isotypes}
\title{Annotate isotypes in V(D)J-seq data}
\usage{
KG_vdj_isotypes(bcr_data, c_call = "c_call")
}
\arguments{
\item{bcr_data}{A data frame containing the BCR data from V(D)J-seq.}

\item{c_call}{Name of the column where C calls are located.}
}
\description{
This function adds a column in a data frame with V(D)J-seq data for isotypes absed on contents in the c_call column.
}
\examples{
KG_vdj_isotypes(BCR_object)
}
