% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_filter_pbmcs.R
\name{KG_filter_pbmcs}
\alias{KG_filter_pbmcs}
\title{A function for differentiating B cells, T cells, and monocytes in PBMC scRNA-seq data.}
\usage{
KG_filter_pbmcs(
  seurat_object,
  features = c("CD19", "CD3E", "CD14"),
  dims = c(1, 2),
  assay = "RNA",
  ncol = 1,
  plot = "violin",
  normalize = TRUE,
  scale.factor = 10000,
  scale = TRUE,
  find.var.feats = FALSE,
  n.var.feats = 2000,
  runPCA = FALSE,
  assay.reproc = "RNA"
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{features}{The genes to use for cell type determination.}

\item{dims}{Which dimensions to plot in the dim plot and potential feature plots.}

\item{ncol}{The number of columns for violin plots and feature plots.}

\item{plot}{Which type of plot to show the features as. Supported options are "violin" and "feature".}

\item{normalize}{Logical; whether data should be normalized after subsetting.}

\item{scale.factor}{Scale factor to use for normalization following filtering.}

\item{scale}{Logical; whether data should be scaled after subsetting.}

\item{find.var.feats}{Logical; whetehr new variable features should be identified after subsetting data.}

\item{n.var.feats}{How many variable features to identify.}

\item{runPCA}{Logical; whether Seurat::RunPCA should be applied after subsetting.}

\item{assay.reproc}{Name of the assay to use when re-processing.}
}
\description{
A function for differentiating B cells, T cells, and monocytes in PBMC scRNA-seq data.
}
\examples{
KG_filter_pbmcs(B_cell_dataset)
}
