% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_get_trajectory.R
\name{KG_get_trajectory}
\alias{KG_get_trajectory}
\title{Extract trajectory}
\usage{
KG_get_trajectory(monocle_object, reduction = "UMAP", is_3D = FALSE)
}
\arguments{
\item{monocle_object}{A monocle3 cell_data_set object.}

\item{reduction}{The dimensionality reduction to use.}

\item{is_3D}{Whether trajectory has been drawn for a 3D dimension plot.}
}
\description{
This function extracts all corrdinates for a trajectory created in monocle3. .
}
\examples{
KG_get_trajectory(cds)
}
