% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_heatmap.R
\name{KG_heatmap}
\alias{KG_heatmap}
\title{Bulk heatmap}
\usage{
KG_heatmap(
  seurat_object,
  genes,
  assay = "RNA",
  group_by = NULL,
  show_genes = TRUE,
  cols = c("#FF00FF", "black", "#FFFF00"),
  heading = TRUE,
  title = NA,
  group_bar_fill_cols = NULL,
  border_col = "black",
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  split_row_at = NULL,
  split_column_at = NULL,
  raster = FALSE
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{genes}{A vector of genes for which to plot the expression.}

\item{assay}{The name of the assay to extract expression values from.}

\item{group_by}{Group variable, if NULL, uses the active identity.}

\item{show_genes}{Whether or not to show the gene names on the y-axis.}

\item{cols}{A vector of colors used for plotting the relative expression.}

\item{heading}{Whether or not to include a heaing to the heatmap.}

\item{title}{The title, if heading = TRUE.}

\item{group_bar_fill_cols}{A vector of colors to use for the group bars. If NULL, default colors are used.}

\item{border_col}{The color for the border of the tiles.}

\item{cluster_rows}{Whether or not genes should be clustered.}

\item{cluster_columns}{Whether or not groups should be clustered.}

\item{split_row_at}{A value, or vector of values, detailing at which row numbers a gap should be included. If NULL, no gaps.}

\item{split_column_at}{A value, or vector of values, detailing at which column numbers a gap should be included. If NULL, no gaps.}

\item{raster}{Whether or not the heatmap should be rasterized.}
}
\description{
This functions draws a heatmap displaying the relative cluster-average expression of specific genes.
}
\examples{
# KG_heatmap(B_cell_dataset, genes = c("CD19", "CD27", "HLA-DRB1"))
}
